python/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pyserial==3.5
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
